Dependabot configuration and security fixes (#169)
* fixes: for dependabot codescanning issues.

* added github workflow configurations for codeql and dependency-review

* fixes: update namespace tests and configuration.

* update go mod

* update go dependecies.

* update go dependecies.

* fixed upgrade issues with gitea library.

* misc fixes to support harvesting from WikiBase sparql endpoint. (#168)

* WIP: misc fixes to support harvesting from WikiBase sparql endpoint.

* update .gitignore

* Cleanup the test and added wikibase test data.

* Feature/mets logging (#164)

* HUB-450: added sublogger functionality

* HUB-226: added source to DaoConfig

* defaultimagepath

* no-return

* no-return-2

* HUB-470: http response code 200 instead of 500 for deleting mets endpoint

* file gets deleted even when not in elasticsearch

* delete mets with source path

* delete mets on sourcepath +

* return error in case of file not found

* HUB-486 start cannot be greater than end

* end = 0 om het einde op te vragen

* review comments

+++ hub3/server/http/handlers/search.go
@@ -283,7 +283,7 @@ func ProcessSearchRequest(w http.ResponseWriter, r *http.Request, searchRequest 
 	w.Header().Add("P_ROWS", strconv.Itoa(int(pager.Rows)))
 
 	// workaround warmer issue ES
-	if res.Hits == nil && pager.Cursor < int32(pager.Total) {
+	if res.Hits == nil && int64(pager.Cursor) < pager.Total {
 		log.Printf("bad response from ES retrying the request")
 		time.Sleep(1 * time.Second)
 		retryCount := r.Context().Value(retryKey)
